requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
openai==1.54.3
pillow==11.0.0
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
soupsieve==2.6
textual==0.85.2
tqdm==4.67.0
trogon==0.6.0
typing_extensions==4.12.2
uc-micro-py==1.0.3
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
